Started using Intellij instead of Eclipse and restuctured project

+++ FireRPG/src/main/java/com/mtgames/firerpg/debug/Console.java
@@ -32,9 +32,9 @@ public class Console extends WindowAdapter implements WindowListener, ActionList
 		// create all components and add them
 		frame = new JFrame("Java Console");
 		Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
-		Dimension frameSize = new Dimension((int) (screenSize.width / 4), (int) (screenSize.height / 2));
-		int x = (int) (frameSize.width / 2 + screenSize.width / 2);
-		int y = (int) (frameSize.height / 2);
+		Dimension frameSize = new Dimension(screenSize.width / 4, screenSize.height / 2);
+		int x = frameSize.width / 2 + screenSize.width / 2;
+		int y = frameSize.height / 2;
 		frame.setBounds(x, y, frameSize.width, frameSize.height);
 		
 		textArea = new JTextArea();
